requirements.txt
Django>=3.2.5,<3.3
psycopg2==2.9.5
import-export==0.2.67.dev6
django-import-export #==2.7.1
xlrd2>=1.3.4,<1.4
openpyxl>=3.0.9,<3.1
requests>=2.27.1,<3
dhis2.py>=2.3.0,<2.4
# django-celery-beat #>=2.2.1,<2.3
# django-celery-results #>=2.2.0,<2.3
# redis>=3.5.3,<3.6
# #uWSGI>=2.0.19.1,<2.1
# celery>=5.1.2,<5.2
django-storages>=1.13.2,<1.14
boto3>=1.26.133,<1.27
Pillow==9.5.0
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.13.0,<4
graphene-django>=2.15.0,<2.16
graphene-file-upload>=1.3.0,<1.4
django-graphql-jwt>=0.3.4,<0.4
PyJWT==2.0.0